Moved unit tests for dirtree.
